Link autosuspend to Binder lifetime
When the system_server process exits for any reason, autosuspend should
be disabled to prevent the device suspending before the framework is
able to restart.
This change adds a IBinder parameter to the enableAutosuspend method,
which is used to check the liveness of the caller. If that Binder
token dies, autosuspend will be disabled until enableAutosuspend() is
called by the framework again.

Bug: 193776617
Test: atest SystemSuspendV1_0UnitTest SystemSuspendV1_0AidlTest
Test: Put device to sleep to enable autosuspend,
      stop framework with `adb shell stop`, `adb shell pkill
      system_server`, or intentionally induced crashes.
      Verify autosuspend gets disabled, and reenabled after framework is
      up and put to sleep again.
Change-Id: If085e3404904fb338d272ec3355f15709122a3ef