system_suspend: use "self" instead of getCallingPid() while reopening a file using fd
While finding filename to re-open it using /proc/<pid>/fd/<fd> mechanism,
we need pid of the process and fd associated with that. Previous solution
was using getCallingPid() to get the pid of the process, which might not
be super accurate, as it would give the pid of the process that requested
the current binder transaction from system suspend, and hence the fd
would point to some other file, safe is use to use "self" instead to
assure that the same file is being referred.

Bug: 192813633
Test: SystemSuspendV1_0UnitTest

Signed-off-by: Manish Varma <varmam@google.com>
Change-Id: Ifd1cf383723bad35c0f2f20f0d4f16ca272d9c34